Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/IndexedDB/transaction-relaxed-durability.tentative.any.worker-expected.txt


FAIL Committed data can be read back out: case 0 assert_equals: expected (string) "default" but got (undefined) undefined
FAIL Committed data can be read back out: case 1 assert_equals: expected (string) "default" but got (undefined) undefined
FAIL Committed data can be read back out: case 2 assert_equals: expected (string) "default" but got (undefined) undefined
FAIL Committed data can be read back out: case 3 assert_equals: expected (string) "relaxed" but got (undefined) undefined
FAIL Committed data can be read back out: case 4 assert_equals: expected (string) "strict" but got (undefined) undefined
FAIL Invalid durability option throws a TypeError assert_throws_js: function "function () {
      db.transaction(['books'], 'readwrite', { durability: 'invalid' });
  }" did not throw
